// embedded-components/src/api/generated/smbdo.schemas.ts
export type AnyValuesMatch = string;

export type QuestionResponseSubQuestionsItem = {
  anyValuesMatch?: AnyValuesMatch;
  questionIds?: QuestionIdList;
};

export type ResponseSchemaType =
  (typeof ResponseSchemaType)[keyof typeof ResponseSchemaType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ResponseSchemaType = {
  array: 'array',
} as const;

/**
 * A subset of JSON Schema used to validate the response values.
 */
export interface ResponseSchema {
  items?: ResponseSchemaItem;
  maxItems?: number;
  minItems?: number;
  type?: ResponseSchemaType;
}
// ...
export interface ResponseSchemaItem {
  exclusiveMaximum?: boolean;
  exclusiveMinimum?: boolean;
  /** Only applicable to string, number, and integer. */
  format?: ResponseSchemaItemFormat;
  maximum?: number;
  maxLength?: number;
  minimum?: number;
  minLength?: number;
  /** Only applicable to string. */
  pattern?: string;
  type?: ResponseSchemaItemType;
}

/**
 * A description of the question.
 */
export type QuestionDescription = string;

/**
 * The identifier for a language and location.
 */
export type LocaleId = string;

/**
 * Content displayed to a user for a given locale.
 */
export interface ContentItem {
  description?: string;
  /** Exact text to be displayed to a user. */
  label: string;
  locale: LocaleId;
}

export interface QuestionResponse {
  /**
   * @minItems 1
   * @maxItems 50
   */
  content?: ContentItem[];
  defaultLocale?: LocaleId;
  description?: QuestionDescription;
  id?: QuestionId;
  parentQuestionId?: QuestionId;
  responseSchema?: ResponseSchema;
  /** @minItems 0 */
  subQuestions?: QuestionResponseSubQuestionsItem[];
}

export interface ClientVerificationResponse {
  /** The timestamp when the request to start verification was accepted. */
  acceptedAt?: string;
  consumerDevice?: ConsumerDevice;
}

export interface ClientVerificationRequest {
  consumerDevice?: ConsumerDevice;
}

/**
 * Describes which attestation to remove. An existing attestation with a matching `documentId` will be removed.
 */
export interface AttestationRemoval {
  documentId: DocumentId;
}

/**
 * Update a party inline within the clients API by supplying its ID.
 */
export interface UpdatePartyRequestInline {
  access?: AccessList;
  email?: Email;
  id?: PartyId;
  individualDetails?: IndividualDetails;
  organizationDetails?: OrganizationDetails;
  /**
   * @minItems 1
   * @maxItems 10
   */
  roles?: Role[];
}

export type UpdateClientRequestSmbdoAddPartiesItem = CreatePartyRequestInline &
  UpdatePartyRequestInline;

export interface UpdateClientRequestSmbdo {
  /**
   * @minItems 1
   * @maxItems 10
   */
  addAttestations?: Attestation[];
  /**
   * Update a party by specifying its ID in the object. Create a party by setting `parentPartyId` to the client's root party. `partyType` and `roles` are required when creating a new party.

   * @minLength 1
   * @maxLength 10
   * @minItems 0
   */
  addParties?: UpdateClientRequestSmbdoAddPartiesItem[];
  addProducts?: ClientProductList;
  consumerDevice?: ConsumerDevice;
  /**
   * @minItems 0
   * @maxItems 200
   */
  questionResponses?: ClientQuestionResponse[];
  /**
   * @minItems 1
   * @maxItems 10
   */
  removeAttestations?: AttestationRemoval[];
}
// ...
export type ResponseValueList = string[];

/**
 * The client's response to a question.
 */
export interface ClientQuestionResponse {
  questionId?: QuestionId;
  values?: ResponseValueList;
}
// ...
export interface CreatePartyRequestInline {
  access?: AccessList;
  email?: Email;
  externalId?: ExternalId;
  individualDetails?: IndividualDetailsRequired;
  organizationDetails?: OrganizationDetailsRequired;
  parentPartyId?: ParentPartyId;
  partyType?: PartyTypeSmbdo;
  /**
   * @minItems 1
   * @maxItems 10
   */
  roles?: Role[];
}
// ...
export interface ConsumerDevice {
  /** IPv4 address of the consumer device. */
  ipAddress?: string;
  /**
   * The unique session identifier of the device that was created by the client. It can be up to 128 characters long and can contain only the following characters: uppercase and lowercase Roman letters, digits, underscore characters, and hyphens (a?z, A?Z, 0?9, _, -). The session ID should contain at least 16 bytes of randomly generated data.

   * @pattern ^[a-zA-Z0-9_-]+$
   */
  sessionId?: string;
}

/**
 * The attesting of presented legal statements. Either "attester" or "attesterFullName" is required.
 */
export interface Attestation {
  attestationTime: string;
  attester?: Attester;
  /** The full name of an individual. */
  attesterFullName?: string;
  documentId: DocumentId;
  /** IPv4 Address. */
  ipAddress: string;
}
// ...
export interface Attester {
  designation: AttesterDesignation;
  firstName: AttesterFirstName;
  lastName: AttesterLastName;
  middleName?: AttesterMiddleName;
}
// ...
export type QuestionId = string;

/**
 * A list of question IDs.
 * @minItems 0
 * @maxItems 200
 */
export type QuestionIdList = QuestionId[];

/**
 * A list of document request IDs.
 * @minItems 0
 * @maxItems 10
 */
export type DocumentRequestIdList = string[];

/**
 * The unique id generated by the system for the uploaded document,  which can be used for future retrieval.

 */
export type DocumentId = string;
// ...
export type ClientProduct = (typeof ClientProduct)[keyof typeof ClientProduct];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ClientProduct = {
  MERCHANT_SERVICES: 'MERCHANT_SERVICES',
  EMBEDDED_PAYMENTS: 'EMBEDDED_PAYMENTS',
} as const;

/**
 * A list of product selections.
 * @minItems 1
 * @maxItems 50
 */
export type ClientProductList = ClientProduct[];